fix : 중복된는 contracts 정리
