+++ core/contracts/curriculum_compose.py
@@ -1,38 +1,10 @@
-from typing import List, Dict, Any, TypedDict, Optional, Literal
-
-class BudgetedTime(TypedDict):
-    total_days: str
-    hours_per_day: str
-    total_hours: str
-
-class UserInfo(TypedDict):
-    purpose: str
-    level: str
-    known_concept: List[str]
-    budgeted_time: BudgetedTime
-    resource_type_preference: List[str]
-
-class CurriculumNode(TypedDict):
-    keyword_id: str
-    keyword: str
-    description: Optional[str]
-    keyword_importance: int
-    is_necessary: Optional[bool]
-    resources: List[Dict[str, Any]]
-
-class CurriculumEdge(TypedDict):
-    start: str
-    end: str
-
-class CurriculumGraphMeta(TypedDict):
-    paper_id: str
-    title: str
-    summarize: str
-
-class CurriculumData(TypedDict):
-    graph_meta: CurriculumGraphMeta
-    nodes: List[CurriculumNode]
-    edges: List[CurriculumEdge]
+from typing import TypedDict, Literal
+from core.contracts.types.user_info import UserInfo
+from core.contracts.types.curriculum import (
+    CurriculumGraph as CurriculumData,
+    KeywordNode as CurriculumNode,
+    KeywordEdge as CurriculumEdge
+)
 
 class CurriculumComposeInput(TypedDict):
     user_info: UserInfo
