+++ core/contracts/study_load_estimation.py
@@ -1,20 +1,10 @@
-from typing import List, Dict, Any, TypedDict
-
-class ResourceData(TypedDict):
-    keyword_id: str
-    keyword: str
-    resource_name: str
-    url: str
-    raw_content: str
-    difficulty: str
-    importance: str
-    study_load: str
-    type: str
+from typing import List, TypedDict
+from core.contracts.types.curriculum import Resource
 
 class StudyLoadEstimationInput(TypedDict):
-    resources: List[ResourceData]
+    resources: List[Resource]
     user_level: str
     purpose: str
 
 class StudyLoadEstimationOutput(TypedDict):
-    evaluated_resources: List[ResourceData]
+    evaluated_resources: List[Resource]
